:tractor: relocate render function to be closer to its Env
